Enhancement: split out behave logic
